Move ContextMenu to the CardBody
The ContextMenu has nothing to do with the main app, it appears in the
`CardBody` or better said DirectoryView. This allows us to later to
consolidate the event listeners and if the ContextMenu should be shown
to the NavigatorCardBody instead of the app itself.

Note: later rename CardBody to DirectoryView.
